backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
sqlalchemy==2.0.35
redis==5.0.8
openai==1.47.1
pyswisseph
python-telegram-bot>=20.0
python-multipart
